+++ src/main/java/com/tienda/repository/UsuarioRepository.java
@@ -9,6 +9,6 @@ import com.tienda.modelo.Usuarios;
 public interface UsuarioRepository  extends CrudRepository<Usuarios, Long>{
 
 	
-	Usuarios findById(int id);
+	Usuarios findById(long id);
 	Usuarios findByNombre(String nombre);
 }
